[Remake] Remake Complementary Color
