+++ WonderfulColor/Classes/WonderfulColor.swift
@@ -7,18 +7,21 @@
 
 #if os(iOS) || os(tvOS) || os(watchOS)
 import UIKit
-
 public typealias WonderfulColor = UIColor
-
 #elseif os(OSX)
 import AppKit
-
 public typealias WonderfulColor = NSColor
-
 #endif
 
 public extension WonderfulColor {
-    var complementary: Complementary {
-        return Complementary(origin: self)
+    var complementary: WonderfulColor {
+        var hue: CGFloat = 0
+        var saturation: CGFloat = 0
+        var brightness: CGFloat = 0
+        var alpha: CGFloat = 0
+        self.getHue(&hue, saturation: &saturation, brightness: &brightness, alpha: &alpha)
+        let compHue = CGFloat((hue + 0.5) - 1.0)
+        
+        return WonderfulColor(hue: compHue, saturation: saturation, brightness: brightness, alpha: alpha)
     }
 }
